+++ src/main/java/com/example/projetoSD/service/UserFileService.java
@@ -8,9 +8,19 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class UserFileService {
-    private static final String FILE_PATH = "../database/users.dat";
+    private static final String FILE_PATH = "./src/main/java/com/example/projetoSD/database/users.dat";
     
     public void saveUsers(List<User> users) {
+        File file = new File(FILE_PATH);
+        if (!file.exists()) {
+            try {
+                boolean __ = file.createNewFile();
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+            return;
+        }
+        
         try (ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(FILE_PATH))) {
             oos.writeObject(users);
         } catch (IOException e) {
@@ -22,6 +32,12 @@ public class UserFileService {
     public List<User> loadUsers() {
         File file = new File(FILE_PATH);
         if (!file.exists()) {
+            try {
+                boolean __ = file.createNewFile();
+            } catch (IOException e) {
+                System.out.println("Error creating file");
+                e.printStackTrace();
+            }
             return new ArrayList<>();
         }
